Refactor 'appObj' entity into its own file
